Move message type preconditions into a dedicated file
